requirements.txt
airtable-python-wrapper==0.15.0
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
google-cloud-secret-manager==2.2.0
hypothesis==6.3.0
identify==1.5.10
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
mailchimp-marketing==3.0.36
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.8
pathspec==0.8.0
pluggy==0.13.1
premailer==3.7.0
pre-commit==2.9.2
py==1.10.0
pydantic==1.8
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.0
python-dotenv==0.16.0
python-http-client==3.3.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
sendgrid==6.4.8
six==1.15.0
slack-sdk==3.4.2
starkbank-ecdsa==1.1.0
structlog==21.1.0
tenacity==6.3.1
toml==0.10.1
twilio==6.47.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.2.2
Werkzeug==1.0.1
